requirements.txt
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-chroma
langchain-core
chromadb
openai
pypdf